refactor: fixed folder structures
