+++ lib/src/pagination/bidirectional_pagination.dart
@@ -6,36 +6,39 @@
 import 'dart:async';
 
 // Project imports:
-import '../pagination_response.dart';
+import '../response/pagination_response.dart';
 import 'base_pagination_control.dart';
-import 'forward_pagination_control.dart';
 import 'pageable.dart';
 import 'pagination.dart';
+import 'pagination_control.dart';
+import 'pagination_control_type.dart';
 import 'paging_event.dart';
 
-typedef ForwardPaging<D, M extends Pageable>
-    = FutureOr<ForwardPaginationControl> Function(
+typedef Paging<D, M extends Pageable> = FutureOr<PaginationControl> Function(
   PagingEvent<D, M> event,
 );
 
-/// This class is an object representing unidirectional paging.
-class UnidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
-  /// Returns the new instance of [UnidirectionalPagination].
-  const UnidirectionalPagination(
+/// This class is an object representing bidirectional
+/// (Forward, Backward) paging.
+class BidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
+  /// Returns the new instance of [BidirectionalPagination].
+  const BidirectionalPagination(
     super.rootPage,
     this.paging,
     super.flipper,
   );
 
   /// The paging callback
-  final ForwardPaging<D, M> paging;
+  final Paging<D, M> paging;
 
   @override
   String? getNextToken(
     final BasePaginationControl control,
     final Pageable? pageableMeta,
   ) =>
-      pageableMeta?.nextToken;
+      control.type.isForward
+          ? pageableMeta?.nextToken
+          : pageableMeta?.previousToken;
 
   @override
   Future<BasePaginationControl> invokePaging(
@@ -51,5 +54,5 @@ class UnidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
     final Map<String, dynamic> queryParameters,
     final String? nextToken,
   ) =>
-      queryParameters['next_token'] = nextToken;
+      queryParameters['pagination_token'] = nextToken;
 }
