+++ lib/src/pagination/unidirectional_pagination.dart
@@ -6,39 +6,36 @@
 import 'dart:async';
 
 // Project imports:
-import '../pagination_response.dart';
+import '../response/pagination_response.dart';
 import 'base_pagination_control.dart';
+import 'forward_pagination_control.dart';
 import 'pageable.dart';
 import 'pagination.dart';
-import 'pagination_control.dart';
-import 'pagination_control_type.dart';
 import 'paging_event.dart';
 
-typedef Paging<D, M extends Pageable> = FutureOr<PaginationControl> Function(
+typedef ForwardPaging<D, M extends Pageable>
+    = FutureOr<ForwardPaginationControl> Function(
   PagingEvent<D, M> event,
 );
 
-/// This class is an object representing bidirectional
-/// (Forward, Backward) paging.
-class BidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
-  /// Returns the new instance of [BidirectionalPagination].
-  const BidirectionalPagination(
+/// This class is an object representing unidirectional paging.
+class UnidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
+  /// Returns the new instance of [UnidirectionalPagination].
+  const UnidirectionalPagination(
     super.rootPage,
     this.paging,
     super.flipper,
   );
 
   /// The paging callback
-  final Paging<D, M> paging;
+  final ForwardPaging<D, M> paging;
 
   @override
   String? getNextToken(
     final BasePaginationControl control,
     final Pageable? pageableMeta,
   ) =>
-      control.type.isForward
-          ? pageableMeta?.nextToken
-          : pageableMeta?.previousToken;
+      pageableMeta?.nextToken;
 
   @override
   Future<BasePaginationControl> invokePaging(
@@ -54,5 +51,5 @@ class BidirectionalPagination<D, M extends Pageable> extends Pagination<D, M> {
     final Map<String, dynamic> queryParameters,
     final String? nextToken,
   ) =>
-      queryParameters['pagination_token'] = nextToken;
+      queryParameters['next_token'] = nextToken;
 }
