+++ lib/src/service/base_service.dart
@@ -10,17 +10,17 @@ import 'package:http/http.dart' as http;
 import 'package:twitter_api_core/twitter_api_core.dart' as core;
 
 // Project imports:
+import '../pagination/bidirectional_pagination.dart';
+import '../pagination/pageable.dart';
+import '../pagination/pagination_context.dart';
+import '../pagination/unidirectional_pagination.dart';
+import '../response/pagination_response.dart';
+import '../response/response_field.dart';
+import '../response/twitter_response.dart';
 import 'common/data.dart';
 import 'common/includes.dart';
 import 'common/meta.dart';
 import 'common/rate_limit.dart';
-import 'pagination/bidirectional_pagination.dart';
-import 'pagination/pageable.dart';
-import 'pagination/pagination_context.dart';
-import 'pagination/unidirectional_pagination.dart';
-import 'pagination_response.dart';
-import 'response_field.dart';
-import 'twitter_response.dart';
 
 /// The callback function for building data object from response.
 typedef DataBuilder<D extends Data> = D Function(Map<String, Object?> json);
