+++ lib/src/service/lists/lists_service.dart
@@ -6,12 +6,12 @@
 import 'package:twitter_api_core/twitter_api_core.dart' as core;
 
 // Project imports:
+import '../../response/twitter_response.dart';
 import '../base_service.dart';
 import '../tweets/tweet_data.dart';
 import '../tweets/tweet_expansion.dart';
 import '../tweets/tweet_field.dart';
 import '../tweets/tweet_meta.dart';
-import '../twitter_response.dart';
 import '../users/user_data.dart';
 import '../users/user_expansion.dart';
 import '../users/user_field.dart';
