+++ lib/src/service/media/media_service.dart
@@ -11,8 +11,8 @@ import 'package:http/http.dart';
 import 'package:twitter_api_core/twitter_api_core.dart' as core;
 
 // Project imports:
+import '../../response/twitter_response.dart';
 import '../base_media_service.dart';
-import '../twitter_response.dart';
 import 'media_category.dart';
 import 'upload_error.dart';
 import 'upload_event.dart';
