+++ lib/src/service/users/users_service.dart
@@ -6,9 +6,9 @@
 import 'package:twitter_api_core/twitter_api_core.dart' as core;
 
 // Project imports:
+import '../../response/twitter_response.dart';
 import '../base_service.dart';
 import '../tweets/tweet_field.dart';
-import '../twitter_response.dart';
 import 'user_data.dart';
 import 'user_expansion.dart';
 import 'user_field.dart';
