+++ test/src/service/compliance/compliance_service_test.dart
@@ -10,11 +10,11 @@ import 'package:twitter_api_core/src/exception/twitter_exception.dart';
 import 'package:twitter_api_core/src/exception/unauthorized_exception.dart';
 
 // Project imports:
+import 'package:twitter_api_v2/src/response/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/compliance/batch_compliance_data.dart';
 import 'package:twitter_api_v2/src/service/compliance/compliance_service.dart';
 import 'package:twitter_api_v2/src/service/compliance/job_status.dart';
 import 'package:twitter_api_v2/src/service/compliance/job_type.dart';
-import 'package:twitter_api_v2/src/service/twitter_response.dart';
 import '../../../mocks/client_context_stubs.dart' as context;
 
 void main() {
