+++ test/src/service/lists/lists_service_test.dart
@@ -8,12 +8,12 @@ import 'package:twitter_api_core/src/client/client_context.dart';
 import 'package:twitter_api_core/src/client/user_context.dart';
 
 // Project imports:
+import 'package:twitter_api_v2/src/response/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/lists/list_data.dart';
 import 'package:twitter_api_v2/src/service/lists/list_meta.dart';
 import 'package:twitter_api_v2/src/service/lists/lists_service.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_data.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_meta.dart';
-import 'package:twitter_api_v2/src/service/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/users/user_data.dart';
 import 'package:twitter_api_v2/src/service/users/user_meta.dart';
 import '../../../mocks/client_context_stubs.dart' as context;
