+++ test/src/service/spaces/spaces_service_test.dart
@@ -8,6 +8,7 @@ import 'package:twitter_api_core/src/client/client_context.dart';
 import 'package:twitter_api_core/src/client/user_context.dart';
 
 // Project imports:
+import 'package:twitter_api_v2/src/response/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/spaces/space_data.dart';
 import 'package:twitter_api_v2/src/service/spaces/space_meta.dart';
 import 'package:twitter_api_v2/src/service/spaces/space_state.dart';
@@ -15,7 +16,6 @@ import 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
 import 'package:twitter_api_v2/src/service/spaces/topic_data.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_data.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_meta.dart';
-import 'package:twitter_api_v2/src/service/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/users/user_data.dart';
 import '../../../mocks/client_context_stubs.dart' as context;
 import '../common_expectations.dart';
