+++ test/src/service/users/users_service_test.dart
@@ -8,7 +8,7 @@ import 'package:twitter_api_core/src/client/client_context.dart';
 import 'package:twitter_api_core/src/client/user_context.dart';
 
 // Project imports:
-import 'package:twitter_api_v2/src/service/twitter_response.dart';
+import 'package:twitter_api_v2/src/response/twitter_response.dart';
 import 'package:twitter_api_v2/src/service/users/user_data.dart';
 import 'package:twitter_api_v2/src/service/users/user_meta.dart';
 import 'package:twitter_api_v2/src/service/users/users_service.dart';
