+++ test/src/twitter_api_test.dart
@@ -9,13 +9,13 @@ import 'package:twitter_api_core/src/exception/twitter_exception.dart';
 import 'package:twitter_api_core/src/exception/unauthorized_exception.dart';
 
 // Project imports:
+import 'package:twitter_api_v2/src/response/twitter_response.dart';
+import 'package:twitter_api_v2/src/response/twitter_stream_response.dart';
 import 'package:twitter_api_v2/src/service/compliance/compliance_service.dart';
 import 'package:twitter_api_v2/src/service/lists/lists_service.dart';
 import 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweet_data.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
-import 'package:twitter_api_v2/src/service/twitter_response.dart';
-import 'package:twitter_api_v2/src/service/twitter_stream_response.dart';
 import 'package:twitter_api_v2/src/service/users/users_service.dart';
 import 'package:twitter_api_v2/src/twitter_api.dart';
